refactor: #18 Enterを押した際の処理をモジュール化
